Update showArticle method in ArticleController and detailArticle template to restrict access to Delete button

+++ views/templates/detailArticle.php
@@ -31,7 +31,9 @@
                 echo '  <div class="detailComment">';
                 echo '      <h3 class="info">Le ' . Utils::convertDateToFrenchFormat($comment->getDateCreation()) . ", " . Utils::format($comment->getPseudo()) . ' a écrit :</h3>';
                 echo '      <p class="content">' . Utils::format($comment->getContent()) . '</p>';
-                echo '      <a class="submit" href="index.php?action=deleteComment&id=' . $comment->getId() . '" ' . Utils::askConfirmation("Êtes-vous sûr(e) de vouloir supprimer ce commentaire ?") . '>Supprimer</a>';
+                if (!empty($isConnected)) {
+                    echo '  <a class="submit" href="index.php?action=deleteComment&id=' . $comment->getId() . '" ' . Utils::askConfirmation("Êtes-vous sûr(e) de vouloir supprimer ce commentaire ?") . '>Supprimer</a>';
+                }
                 echo '  </div>';
                 echo '</li>';
             }
